Add FlaggedFileAsInRoot test
