+++ ProjFS.Mac/PrjFSKext/KauthHandler.cpp
@@ -10,7 +10,6 @@
 #include "VirtualizationRoots.hpp"
 #include "VnodeUtilities.hpp"
 #include "KauthHandler.hpp"
-#include "KextLog.hpp"
 #include "public/Message.h"
 #include "Locks.hpp"
 #include "PrjFSProviderUserClient.hpp"
@@ -19,6 +18,9 @@
 
 #ifdef KEXT_UNIT_TESTING
 #include "KauthHandlerTestable.hpp"
+#else
+#include "KextLog.hpp"
+#include "KauthHandlerHelper.h"
 #endif
 
 // Function prototypes
@@ -42,9 +44,8 @@ static int HandleFileOpOperation(
 
 static int GetPid(vfs_context_t _Nonnull context);
 
-static bool TryReadVNodeFileFlags(vnode_t vn, vfs_context_t _Nonnull context, uint32_t* flags);
 KEXT_STATIC_INLINE bool FileFlagsBitIsSet(uint32_t fileFlags, uint32_t bit);
-static inline bool TryGetFileIsFlaggedAsInRoot(vnode_t vnode, vfs_context_t _Nonnull context, bool* flaggedInRoot);
+KEXT_STATIC_INLINE bool TryGetFileIsFlaggedAsInRoot(vnode_t vnode, vfs_context_t _Nonnull context, bool* flaggedInRoot);
 KEXT_STATIC_INLINE bool ActionBitIsSet(kauth_action_t action, kauth_action_t mask);
 
 KEXT_STATIC bool IsFileSystemCrawler(const char* procname);
@@ -61,7 +62,7 @@ static bool TrySendRequestAndWaitForResponse(
     int* kauthResult,
     int* kauthError);
 static void AbortAllOutstandingEvents();
-static bool ShouldIgnoreVnodeType(vtype vnodeType, vnode_t vnode);
+KEXT_STATIC bool ShouldIgnoreVnodeType(vtype vnodeType, vnode_t vnode);
 
 static bool ShouldHandleVnodeOpEvent(
     // In params:
@@ -1087,42 +1088,13 @@ static int GetPid(vfs_context_t _Nonnull context)
     return proc_pid(callingProcess);
 }
 
-static errno_t GetVNodeAttributes(vnode_t vn, vfs_context_t _Nonnull context, struct vnode_attr* attrs)
-{
-    VATTR_INIT(attrs);
-    VATTR_WANTED(attrs, va_flags);
-    
-    return vnode_getattr(vn, attrs, context);
-}
-
-static bool TryReadVNodeFileFlags(vnode_t vn, vfs_context_t _Nonnull context, uint32_t* flags)
-{
-    struct vnode_attr attributes = {};
-    *flags = 0;
-    errno_t err = GetVNodeAttributes(vn, context, &attributes);
-    if (0 != err)
-    {
-        // TODO(Mac): May fail on some file system types? Perhaps we should early-out depending on mount point anyway.
-        // We should also consider:
-        //   - Logging this error
-        //   - Falling back on vnode lookup (or custom cache) to determine if file is in the root
-        //   - Assuming files are empty if we can't read the flags
-        KextLog_FileError(vn, "ReadVNodeFileFlags: GetVNodeAttributes failed with error %d; vnode type: %d, recycled: %s", err, vnode_vtype(vn), vnode_isrecycled(vn) ? "yes" : "no");
-        return false;
-    }
-    
-    assert(VATTR_IS_SUPPORTED(&attributes, va_flags));
-    *flags = attributes.va_flags;
-    return true;
-}
-
 KEXT_STATIC_INLINE bool FileFlagsBitIsSet(uint32_t fileFlags, uint32_t bit)
 {
     // Note: if multiple bits are set in 'bit', this will return true if ANY are set in fileFlags
     return 0 != (fileFlags & bit);
 }
 
-static inline bool TryGetFileIsFlaggedAsInRoot(vnode_t vnode, vfs_context_t _Nonnull context, bool* flaggedInRoot)
+KEXT_STATIC_INLINE bool TryGetFileIsFlaggedAsInRoot(vnode_t vnode, vfs_context_t _Nonnull context, bool* flaggedInRoot)
 {
     uint32_t vnodeFileFlags;
     *flaggedInRoot = false;
@@ -1155,7 +1127,7 @@ KEXT_STATIC bool IsFileSystemCrawler(const char* procname)
     return false;
 }
 
-static bool ShouldIgnoreVnodeType(vtype vnodeType, vnode_t vnode)
+KEXT_STATIC bool ShouldIgnoreVnodeType(vtype vnodeType, vnode_t vnode)
 {
     switch (vnodeType)
     {
