+++ ProjFS.Mac/PrjFSKext/KauthHandlerTestable.hpp
@@ -1,5 +1,7 @@
 #include "public/PrjFSCommon.h"
 #include <sys/kernel_types.h>
+#include <sys/vnode.h>
+#include "../PrjFSKextTests/KextLogMock.h"
 
 #ifndef __cplusplus
 #error None of the kext code is set up for being called from C or Objective-C; change the including file to C++ or Objective-C++
@@ -12,4 +14,6 @@
 KEXT_STATIC_INLINE bool FileFlagsBitIsSet(uint32_t fileFlags, uint32_t bit);
 KEXT_STATIC_INLINE bool ActionBitIsSet(kauth_action_t action, kauth_action_t mask);
 KEXT_STATIC bool IsFileSystemCrawler(const char* procname);
-
+KEXT_STATIC bool ShouldIgnoreVnodeType(vtype vnodeType, vnode_t vnode);
+KEXT_STATIC_INLINE bool TryGetFileIsFlaggedAsInRoot(vnode_t vnode, vfs_context_t context, bool* flaggedInRoot);
+bool TryReadVNodeFileFlags(vnode_t vn, vfs_context_t context, uint32_t* flags);
